requirements-dev.txt
-r requirements.txt

# Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==24.2.0

# Code quality
flake8==7.0.0
black==24.3.0
isort==5.13.2
mypy==1.9.0
django-stubs==4.2.7

# Dev utilities
ipython==8.22.2
django-debug-toolbar==4.3.0
